Tidying up code a bit
